task5_search_images_with_image/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
pandas>=2.1.0
scikit-learn>=1.3.0
Pillow>=10.0.0
streamlit>=1.30.0
numpy>=1.26.0